Refactor missing types and progress on migration to types.d.ts
