+++ components/map/legend/item/component.tsx
@@ -1,13 +1,7 @@
-import { FC, ReactNode } from 'react';
+import { FC } from 'react';
 import cx from 'classnames';
 
-export interface LegendItemProps {
-  id: string;
-  name: string;
-  description?: string;
-  icon?: ReactNode;
-  children?: ReactNode;
-}
+import { LegendItemProps } from './types';
 
 export const LegendItem: FC<LegendItemProps> = ({
   id,
