+++ components/map/legend/item/types.d.ts
@@ -1,0 +1,9 @@
+import { ReactNode } from 'react';
+
+export interface LegendItemProps {
+  id: string;
+  name: string;
+  description?: string;
+  icon?: ReactNode;
+  children?: ReactNode;
+}
