+++ components/map/legend/sortable/list/item/component.tsx
@@ -1,13 +1,10 @@
-import { FC, ReactNode } from 'react';
+import { FC } from 'react';
 import cx from 'classnames';
 
 import { useSortable } from '@dnd-kit/sortable';
 import { CSS } from '@dnd-kit/utilities';
 
-export interface SortableItemProps {
-  id: string;
-  children: ReactNode;
-}
+import { SortableItemProps } from './types';
 
 export const SortableItem: FC<SortableItemProps> = ({ id, children }: SortableItemProps) => {
   const { attributes, listeners, setNodeRef, transform, transition, isDragging } = useSortable({
