+++ components/map/legend/sortable/list/item/types.d.ts
@@ -1,0 +1,6 @@
+import { ReactNode } from 'react';
+
+export interface SortableItemProps {
+  id: string;
+  children: ReactNode;
+}
