+++ components/map/legend/types.d.ts
@@ -1,0 +1,6 @@
+export interface LegendProps {
+  className?: string;
+  children: React.ReactNode;
+  maxHeight: string | number;
+  onChangeOrder: (id: string[]) => void;
+}
