+++ components/map/legend/types/basic/types.d.ts
@@ -1,0 +1,7 @@
+export interface LegendTypeBasicProps {
+  className?: string;
+  items: Array<{
+    value: string;
+    color: string;
+  }>;
+}
