+++ components/map/legend/types/choropleth/component.tsx
@@ -1,13 +1,7 @@
 import { FC } from 'react';
 import cx from 'classnames';
 
-export interface LegendTypeChoroplethProps {
-  className?: string;
-  items: Array<{
-    value: string;
-    color: string;
-  }>;
-}
+import { LegendTypeChoroplethProps } from './types';
 
 export const LegendTypeChoropleth: FC<LegendTypeChoroplethProps> = ({
   className = '',
