+++ components/map/legend/types/choropleth/types.d.ts
@@ -1,0 +1,7 @@
+export interface LegendTypeChoroplethProps {
+  className?: string;
+  items: Array<{
+    value: string;
+    color: string;
+  }>;
+}
