+++ components/map/legend/types/gradient/types.d.ts
@@ -1,0 +1,7 @@
+export interface LegendTypeGradientProps {
+  className?: string;
+  items: Array<{
+    value: string;
+    color: string;
+  }>;
+}
