+++ components/search/types.d.ts
