+++ components/tooltip/arrow/types.d.ts
@@ -1,0 +1,3 @@
+export interface ArrowProps {
+  'data-placement'?: string;
+}
